:art:Modify : API 코드 통합 및 수정
