--- src/api/emailDuplicateAPI.js
@@ -1,17 +1,0 @@
-const emailDuplicateAPI = async (userEmail) => {
-  const res = await fetch(`https://mandarin.api.weniv.co.kr/user/emailvalid`, {
-    method: 'POST',
-    headers: {
-      'Content-type': 'application/json',
-    },
-    body: JSON.stringify({
-      user: {
-        email: userEmail,
-      },
-    }),
-  });
-  const json = await res.json();
-  return json;
-};
-
-export default emailDuplicateAPI;
