--- src/api/getFollowersPostsAPI.js
@@ -1,21 +1,0 @@
-const getFollowersPostsAPI = async (token) => {
-  try {
-    const res = await fetch(
-      `https://mandarin.api.weniv.co.kr/post/feed/?limit=Number&skip=Number`,
-      {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'Content-type': 'application/json',
-        },
-      },
-    );
-    const json = await res.json();
-    return json.posts;
-  } catch (error) {
-    console.log(error);
-    return null;
-  }
-};
-
-export default getFollowersPostsAPI;
