--- src/api/unlikeAPI.js
@@ -1,24 +1,0 @@
-const unlikeAPI = async (postId, token) => {
-  try {
-    const res = await fetch(
-      `https://mandarin.api.weniv.co.kr/post/${postId}/unheart`,
-      {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'Content-type': 'application/json',
-        },
-      },
-    );
-    const json = await res.json();
-    if (!res.ok) {
-      // eslint-disable-next-line no-alert
-      alert(`${json.message}`);
-    }
-    return json.post;
-  } catch (error) {
-    console.log(error);
-    return null;
-  }
-};
-export default unlikeAPI;
